backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
python-multipart==0.0.26
python-dotenv==1.2.2
pydantic==2.12.5
groq==1.2.0
PyMuPDF==1.27.2.2
scikit-learn==1.8.0
numpy==2.4.4
requests==2.33.1
